requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
playwright==1.41.0
google-generativeai==0.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
PyPDF2==3.0.1
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1
Pillow==10.2.0
beautifulsoup4==4.12.3
lxml==5.1.0
openpyxl==3.1.2
python-multipart==0.0.6
pydantic==2.5.3
opencv-python-headless==4.9.0.80
pydub==0.25.1
moviepy==1.0.3
python-magic==0.4.27
filetype==1.2.0